pyproject.toml
[project]
name = "rss-aggregator"
version = "0.8.0"
description = "FastAPI-based RSS feed aggregator with source management, scheduled fetching, and filtering"
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.12"
authors = [
    { name = "Kimhsiao", email = "white.shopping@gmail.com" }
]
dependencies = [
    "aiosqlite",
    "alembic",
    "fastapi",
    "feedparser",
    "greenlet>=3.3.2",
    "httpx",
    "pydantic-settings",
    "python-multipart",
    "sqlalchemy",
    "uvicorn[standard]",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "mypy>=1.19.1",
    "pyinstaller>=6.0.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.6",
]
